+++ src/base/Plugin.php
@@ -126,21 +126,15 @@ class Plugin extends Module implements PluginInterface
     /**
      * @inheritdoc
      */
-    public function install()
+    public function install(): void
     {
-        if ($this->beforeInstall() === false) {
-            return false;
-        }
+        $this->beforeInstall();
 
         $migrator = $this->getMigrator();
 
         // Run the install migration, if there is one
         if (($migration = $this->createInstallMigration()) !== null) {
-            try {
-                $migrator->migrateUp($migration);
-            } catch (MigrationException $e) {
-                return false;
-            }
+            $migrator->migrateUp($migration);
         }
 
         // Mark all existing migrations as applied
@@ -151,36 +145,26 @@ class Plugin extends Module implements PluginInterface
         $this->isInstalled = true;
 
         $this->afterInstall();
-
-        return null;
     }
 
     /**
      * @inheritdoc
      */
-    public function uninstall()
+    public function uninstall(): void
     {
-        if ($this->beforeUninstall() === false) {
-            return false;
-        }
+        $this->beforeUninstall();
 
         if (($migration = $this->createInstallMigration()) !== null) {
-            try {
-                $this->getMigrator()->migrateDown($migration);
-            } catch (MigrationException $e) {
-                return false;
-            }
+            $this->getMigrator()->migrateDown($migration);
         }
 
         $this->afterUninstall();
-
-        return null;
     }
 
     /**
      * @inheritdoc
      */
-    public function getSettings()
+    public function getSettings(): ?Model
     {
         if ($this->_settingsModel === null) {
             $this->_settingsModel = $this->createSettingsModel() ?: false;
@@ -196,7 +180,7 @@ class Plugin extends Module implements PluginInterface
     /**
      * @inheritdoc
      */
-    public function setSettings(array $settings)
+    public function setSettings(array $settings): void
     {
         if (($model = $this->getSettings()) === null) {
             Craft::warning('Attempting to set settings on a plugin that doesn\'t have settings: ' . $this->id);
@@ -237,7 +221,7 @@ class Plugin extends Module implements PluginInterface
     /**
      * @inheritdoc
      */
-    public function getCpNavItem()
+    public function getCpNavItem(): ?array
     {
         $ret = [
             'label' => $this->name,
@@ -320,7 +304,7 @@ class Plugin extends Module implements PluginInterface
     /**
      * @inheritdoc
      */
-    public function afterSaveSettings()
+    public function afterSaveSettings(): void
     {
         // Trigger an 'afterSaveSettings' event
         if ($this->hasEventHandlers(self::EVENT_AFTER_SAVE_SETTINGS)) {
@@ -352,34 +336,36 @@ class Plugin extends Module implements PluginInterface
     /**
      * Performs actions before the plugin is installed.
      *
-     * @return bool Whether the plugin should be installed
+     * @return void
      */
-    protected function beforeInstall(): bool
+    protected function beforeInstall(): void
     {
-        return true;
     }
 
     /**
      * Performs actions after the plugin is installed.
+     *
+     * @return void
      */
-    protected function afterInstall()
+    protected function afterInstall(): void
     {
     }
 
     /**
      * Performs actions before the plugin is uninstalled.
      *
-     * @return bool Whether the plugin should be uninstalled
+     * @return void
      */
-    protected function beforeUninstall(): bool
+    protected function beforeUninstall(): void
     {
-        return true;
     }
 
     /**
      * Performs actions after the plugin is uninstalled.
+     *
+     * @return void
      */
-    protected function afterUninstall()
+    protected function afterUninstall(): void
     {
     }
 
@@ -388,7 +374,7 @@ class Plugin extends Module implements PluginInterface
      *
      * @return Model|null
      */
-    protected function createSettingsModel()
+    protected function createSettingsModel(): ?Model
     {
         return null;
     }