+++ src/base/PluginInterface.php
@@ -47,18 +47,16 @@ interface PluginInterface
     /**
      * Installs the plugin.
      *
-     * @return null|false Return `false` to indicate the installation failed.
-     * All other return values mean the installation was successful.
+     * @return void
      */
-    public function install();
+    public function install(): void;
 
     /**
      * Uninstalls the plugin.
      *
-     * @return null|false Return `false` to indicate the uninstallation failed.
-     * All other return values mean the uninstallation was successful.
+     * @return void
      */
-    public function uninstall();
+    public function uninstall(): void;
 
     /**
      * Returns the plugin’s migration manager
@@ -72,14 +70,15 @@ interface PluginInterface
      *
      * @return Model|null The model that the plugin’s settings should be stored on, if the plugin has settings
      */
-    public function getSettings();
+    public function getSettings(): ?Model;
 
     /**
      * Sets the plugin settings
      *
      * @param array $settings The plugin settings that should be set on the settings model
+     * @return void
      */
-    public function setSettings(array $settings);
+    public function setSettings(array $settings): void;
 
     /**
      * Returns the settings page response.
@@ -129,7 +128,7 @@ interface PluginInterface
      * @see PluginTrait::$hasCpSection
      * @see Cp::nav()
      */
-    public function getCpNavItem();
+    public function getCpNavItem(): ?array;
 
     // Events
     // -------------------------------------------------------------------------
@@ -145,7 +144,8 @@ interface PluginInterface
     /**
      * Performs actions after the plugin’s settings are saved.
      *
+     * @return void
      * @since 3.0.16
      */
-    public function afterSaveSettings();
+    public function afterSaveSettings(): void;
 }